refactor(stringAdder): 폴더 변경
