+++ src/test/java/stringadder/domain/AdderTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import stringadder.utils.exception.StringUtils;
+import stringadder.utils.StringUtils;
 
 public class AdderTest {
